+++ src/strategy/unions.rs
@@ -16,12 +16,16 @@ use rand::distributions::IndependentSample;
 use strategy::traits::*;
 use test_runner::*;
 
+/// A **relative** `weight` of a particular `Strategy` corresponding to `T`
+/// coupled with `T` itself. The weight is currently given in `u32`.
+pub type W<T> = (u32, T);
+
 /// A `Strategy` which picks from one of several delegate `Stragegy`s.
 ///
 /// See `Strategy::prop_union()`.
 #[derive(Clone, Debug)]
 pub struct Union<T : Strategy> {
-    options: Vec<(u32,T)>,
+    options: Vec<W<T>>,
 }
 
 impl<T : Strategy> Union<T> {
@@ -36,9 +40,9 @@ impl<T : Strategy> Union<T> {
     ///
     /// Panics if `options` is empty.
     pub fn new<I : IntoIterator<Item = T>>(options: I) -> Self {
-        let options: Vec<(u32,T)> = options.into_iter()
+        let options: Vec<W<T>> = options.into_iter()
             .map(|v| (1, v)).collect();
-        assert!(options.len() > 0);
+        assert!(!options.is_empty());
 
         Union { options }
     }
@@ -55,12 +59,12 @@ impl<T : Strategy> Union<T> {
     /// Panics if `options` is empty or any element has a weight of 0.
     ///
     /// Panics if the sum of the weights overflows a `u32`.
-    pub fn new_weighted(options: Vec<(u32,T)>) -> Self {
-        assert!(options.len() > 0);
+    pub fn new_weighted(options: Vec<W<T>>) -> Self {
+        assert!(!options.is_empty());
         assert!(!options.iter().any(|&(w, _)| 0 == w),
                 "Union option has a weight of 0");
-        assert!(options.iter().map(|&(w, _)| w as u64).sum::<u64>() <=
-                u32::MAX as u64, "Union weights overflow u32");
+        assert!(options.iter().map(|&(w, _)| u64::from(w)).sum::<u64>() <=
+                u64::from(u32::MAX), "Union weights overflow u32");
         Union { options }
     }
 
@@ -87,7 +91,7 @@ impl<T : Strategy> Strategy for Union<T> {
 
     fn new_value(&self, runner: &mut TestRunner)
                  -> Result<Self::Value, String> {
-        fn extract_weight<V>(&(w, _): &(u32, V)) -> u32 { w }
+        fn extract_weight<V>(&(w, _): &W<V>) -> u32 { w }
 
         let pick = pick_weighted(
             runner,
@@ -216,6 +220,7 @@ impl<T> TupleUnion<T> {
     /// must be a 2- to 10-tuple of `(u32, impl Strategy)` pairs where all
     /// strategies ultimately produce the same value. Each `u32` indicates the
     /// relative weight of its corresponding strategy.
+    /// You may use `W<S>` as an alias for `(u32, S>`.
     ///
     /// Using this constructor directly is discouraged; prefer to use
     /// `prop_oneof!` since it is generally clearer.
@@ -227,9 +232,8 @@ impl<T> TupleUnion<T> {
 macro_rules! tuple_union {
     ($($gen:ident $ix:tt)*) => {
         impl<A : Strategy, $($gen: Strategy),*> Strategy
-        for TupleUnion<((u32, A), $((u32, $gen)),*)>
-        where $($gen::Value : ValueTree<Value =
-                <A::Value as ValueTree>::Value>),* {
+        for TupleUnion<(W<A>, $(W<$gen>),*)>
+        where $($gen::Value : ValueTree<Value = ValueFor<A>>),* {
             type Value = TupleUnionValueTree<
                 (A::Value, $(Option<$gen::Value>),*)>;
 
@@ -294,7 +298,7 @@ value_tree_tuple!(access_tuple8, B C D E F G H);
 value_tree_tuple!(access_tuple9, B C D E F G H I);
 value_tree_tuple!(access_tupleA, B C D E F G H I J);
 
-const WEIGHT_BASE: u32 = 0x80000000;
+const WEIGHT_BASE: u32 = 0x8000_0000;
 
 /// Convert a floating-point weight in the range (0.0,1.0) to a pair of weights
 /// that can be used with `Union` and similar.
@@ -315,7 +319,7 @@ pub fn float_to_weight(f: f64) -> (u32, u32) {
 
     // Clamp to 1..WEIGHT_BASE-1 so that we never produce a weight of 0.
     let pos = max(1, min(WEIGHT_BASE - 1,
-                         (f * WEIGHT_BASE as f64).round() as u32));
+                         (f * f64::from(WEIGHT_BASE)).round() as u32));
     let neg = WEIGHT_BASE - pos;
 
     (pos, neg)